Extract events crate inline tests to sibling files
Move channel_bus, event, and subscription test modules into
*_tests.rs siblings via #[path] so production files stay small
without changing test paths or behavior.
